+++ src/debugger/commands/commands/command_debug_enter.ts
@@ -1,0 +1,9 @@
+import { Command } from "../command";
+import { Mediator } from "../../mediator";
+
+export class CommandDebugEnter extends Command {
+	public trigger(parameters: any[]) {
+		let reason: string = parameters[1];
+		Mediator.notify("debug_enter", [reason]);
+	}
+}