+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### 1.1
+* Add the debugger to the extension
+
+### 1.0.3
+* Fix hover popup position for VSCode 1.42+
+
 ### 1.0.1
 * Fix run editor error on windows with default terminal configurations